move test_bird to test_clearsky and move bird to clearsky core module
